feat: WIP added some repository tests but they are failing

+++ services/src/main/java/ar/edu/itba/paw/services/UserServiceImpl.java
@@ -16,7 +16,7 @@ public class UserServiceImpl implements UserService {
     private UserDao userDao;
 
     @Override
-    public Optional<User> findById(Integer id) {
+    public Optional<User> findById(Long id) {
         return this.userDao.findById(id);
     }
